Win32 Fixes

+++ include/aws/s3/private/s3_meta_request_impl.h
@@ -187,7 +187,7 @@ struct aws_s3_meta_request {
     struct aws_http_message *initial_request_message;
 
     /* Part size to use for uploads and downloads.  Passed down by the creating client. */
-    const uint64_t part_size;
+    const size_t part_size;
 
     struct aws_cached_signing_config_aws *cached_signing_config;
 
@@ -250,14 +250,14 @@ struct aws_s3_meta_request {
 struct aws_s3_meta_request *aws_s3_meta_request_auto_ranged_get_new(
     struct aws_allocator *allocator,
     struct aws_s3_client *client,
-    uint64_t part_size,
+    size_t part_size,
     const struct aws_s3_meta_request_options *options);
 
 /* Creates a new auto-ranged put meta request.  This will do a multipart upload in parallel when appropriate. */
 struct aws_s3_meta_request *aws_s3_meta_request_auto_ranged_put_new(
     struct aws_allocator *allocator,
     struct aws_s3_client *client,
-    uint64_t part_size,
+    size_t part_size,
     uint32_t num_parts,
     const struct aws_s3_meta_request_options *options);
 
@@ -286,7 +286,7 @@ AWS_S3_API
 int aws_s3_meta_request_init_base(
     struct aws_allocator *allocator,
     struct aws_s3_client *client,
-    uint64_t part_size,
+    size_t part_size,
     const struct aws_s3_meta_request_options *options,
     void *impl,
     struct aws_s3_meta_request_vtable *vtable,
